cfg-block-generator: add reference argument
The new "reference" argument can be used to refer back to the spot in the
configuration that caused the block to be pasted in. It contains the
name of the config file, line and column information where the inclusion
is coming from.

It is to be used in error messages.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>